commerce-gateway/pyproject.toml
[project]
name = "ozon-pilot-commerce-gateway"
version = "0.1.0"
description = "Commercial API gateway for OZON-PILOT business adapters."
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.13.2",
    "asyncpg>=0.29.0",
    "bcrypt>=4.0,<4.1",
    "email-validator>=2.2.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.2",
    "jinja2>=3.1.4",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.4.0",
    "pyjwt>=2.9.0",
    "python-multipart>=0.0.9",
    "redis>=5.0.8",
    "sqlalchemy[asyncio]>=2.0.32",
    "uvicorn[standard]>=0.30.6",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.8",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["migrations*", "tests*"]
